未設定一覧ページからボタンを削除

+++ resources/views/layouts/folder_index_template.blade.php
@@ -49,10 +49,12 @@
     {{-- フォルダ名表示 --}}
     <h3 class="pt-3">{{ $fold->name }}</h3>
     <div class="py-3">
-      {{-- フォルダへToDo追加ボタン --}}
-      <a href="/add_folder_form/{{ $fold->id }}" class="btn btn-primary">フォルダへ追加</></a>
-      {{-- フォルダ削除ボタン --}}
-      <a href="/delete_folder_confirm/{{ $fold->id }}" class="btn btn-danger">フォルダを削除</></a>
+      @if($fold->id != 0)
+        {{-- フォルダへToDo追加ボタン --}}
+        <a href="/add_folder_form/{{ $fold->id }}" class="btn btn-primary">フォルダへ追加</></a>
+        {{-- フォルダ削除ボタン --}}
+        <a href="/delete_folder_confirm/{{ $fold->id }}" class="btn btn-danger">フォルダを削除</></a>
+      @endif
       {{-- 検索ワードと結果件数表示 --}}
       @if($search)
         <a class="pl-2 text-muted">検索ワード：{{ $search }}</a>
